+++ app/src/main/java/com/clloret/days/events/common/CommonEventView.java
@@ -1,0 +1,16 @@
+package com.clloret.days.events.common;
+
+public interface CommonEventView {
+
+  void showPeriodText(String text);
+
+  void showDate(String text);
+
+  void showSelectedTags(String text);
+
+  void showSelectedReminder(String text);
+
+  void showSelectedTimeLapseReset(String text);
+
+  void showError(String text);
+}