+++ app/src/main/java/com/clloret/days/events/common/SelectTagsHelper.java
@@ -2,9 +2,11 @@ package com.clloret.days.events.common;
 
 import android.content.res.Resources;
 import android.text.TextUtils;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentManager;
 import com.clloret.days.R;
-import com.clloret.days.base.BaseMvpActivity;
 import com.clloret.days.domain.utils.SelectionMap;
+import com.clloret.days.events.common.SelectTagsDialog.SelectTagsDialogListener;
 import com.clloret.days.model.entities.EventViewModel;
 import com.clloret.days.model.entities.TagViewModel;
 import io.reactivex.Observable;
@@ -17,6 +19,7 @@ import javax.inject.Inject;
 public class SelectTagsHelper {
 
   private final Resources resources;
+  private SelectTagsDialogListener dialogListener;
   private SelectionMap<String, TagViewModel> mapTags = new SelectionMap<>();
 
   @Inject
@@ -65,7 +68,8 @@ public class SelectTagsHelper {
     }
   }
 
-  public void showSelectTagsDialog(BaseMvpActivity activity, SelectTagsHelperListener listener) {
+  public void showSelectTagsDialog(FragmentManager fragmentManager,
+      SelectTagsHelperListener listener) {
 
     if (mapTags.size() == 0) {
       listener.onError(resources.getString(R.string.msg_error_no_tags_available));
@@ -96,8 +100,8 @@ public class SelectTagsHelper {
     SelectTagsDialog dialog = SelectTagsDialog
         .newInstance(resources.getString(R.string.title_select_tags),
             nameTags.toArray(new String[0]), checkedTags,
-            new ArrayList<>(tags));
-    dialog.show(activity.getSupportFragmentManager(), "tags");
+            new ArrayList<>(tags), dialogListener);
+    dialog.show(fragmentManager, "tags");
   }
 
   public void updateSelectedTags(Collection<TagViewModel> selectedItems) {
@@ -114,6 +118,11 @@ public class SelectTagsHelper {
   public void addTagToSelection(TagViewModel tag) {
 
     mapTags.addToSelection(tag);
+  }
+
+  public void setDialogListener(@Nullable SelectTagsDialogListener dialogListener) {
+
+    this.dialogListener = dialogListener;
   }
 
   public interface SelectTagsHelperListener {